issue #987 向け STTR1 入出力回帰テストを追加
