+++ tests/tbx_lib_tests.rs
@@ -20,30 +20,6 @@ fn run_tbx_test(path: &PathBuf, base_dir: &Path) -> Result<(), String> {
         .map_err(|e| format!("{}: {e}", path.display()))
 }
 
-#[derive(Clone, Default)]
-struct SharedOutput(std::sync::Arc<std::sync::Mutex<Vec<u8>>>);
-
-impl SharedOutput {
-    fn into_string(self) -> String {
-        let bytes = self.0.lock().expect("shared output lock poisoned").clone();
-        String::from_utf8(bytes).expect("shared output must be valid UTF-8")
-    }
-}
-
-impl std::io::Write for SharedOutput {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        self.0
-            .lock()
-            .expect("shared output lock poisoned")
-            .extend_from_slice(buf);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-
 // Individual #[test] functions for each lib/tests/test_*.tbx file are generated
 // by build.rs and included here.
 include!(concat!(env!("OUT_DIR"), "/tbx_lib_tests_generated.rs"));
@@ -2028,233 +2004,6 @@ fn test_getdec_safe_partial_numeric_tuple_projection() {
     assert_eq!(
         out, "0 FALSE",
         "partial numeric input should produce TUPLE(0, FALSE)"
-    );
-}
-
-fn run_trek_src_with_input_and_flushed_output(tbx_src: &str, mock_input: &str) -> (String, String) {
-    use std::io::Cursor;
-
-    let mut interp = Interpreter::new();
-    interp
-        .set_base_dir(PathBuf::from(env!("CARGO_MANIFEST_DIR")))
-        .expect("manifest dir is absolute");
-    interp.vm_mut().input_reader = Box::new(Cursor::new(mock_input.to_string()));
-    let flushed = SharedOutput::default();
-    interp.vm_mut().output_writer = Box::new(flushed.clone());
-    interp
-        .exec_source(tbx_src)
-        .unwrap_or_else(|e| panic!("exec_source failed: {e}"));
-    (interp.take_output(), flushed.into_string())
-}
-
-#[test]
-fn test_getdec_safe_test_mode_keeps_prompt_in_output_buffer() {
-    let src = concat!(
-        "DEF CHECK()\n",
-        "  PUTSTR \"COMMAND:\"\n",
-        "  VAR CMD = GETDEC?()\n",
-        "  PUTSTR GET_OUTPUT()\n",
-        "  PUTSTR \" \"\n",
-        "  PUTVAL CMD[2]\n",
-        "END\n",
-        "CHECK\n",
-    );
-    let out = run_with_input_and_flush_mode(src, "123\n", InputFlushMode::KeepBufferedForTest);
-    assert_eq!(out, "COMMAND: TRUE");
-}
-
-#[test]
-fn test_library_computer_eof_flushes_prompt_and_buffers_menu() {
-    let src = concat!(
-        "USE \"examples/trek/state.tbx\"\n",
-        "USE \"examples/trek/util.tbx\"\n",
-        "USE \"examples/trek/init.tbx\"\n",
-        "USE \"examples/trek/scan.tbx\"\n",
-        "USE \"examples/trek/combat.tbx\"\n",
-        "USE \"examples/trek/nav.tbx\"\n",
-        "USE \"examples/trek/library.tbx\"\n",
-        "DEF RUN()\n",
-        "  INIT_GAME\n",
-        "  VAR DISCARDED = GET_OUTPUT()\n",
-        "  LIBRARY_COMPUTER\n",
-        "END\n",
-        "RUN\n",
-    );
-    let (buffered, flushed) = run_trek_src_with_input_and_flushed_output(src, "");
-    assert_eq!(flushed, "COMPUTER ACTIVE AND AWAITING COMMAND");
-    assert_eq!(
-        buffered,
-        concat!(
-            "FUNCTIONS AVAILABLE FROM COMPUTER\n",
-            "   0 = CUMULATIVE GALACTIC RECORD\n",
-            "   1 = STATUS REPORT\n",
-            "   2 = PHOTON TORPEDO DATA\n",
-        )
-    );
-}
-
-#[test]
-fn test_trek_command_loop_refreshes_docking_after_navigation_before_next_prompt() {
-    let src = concat!(
-        "USE \"examples/trek/state.tbx\"\n",
-        "USE \"examples/trek/util.tbx\"\n",
-        "USE \"examples/trek/init.tbx\"\n",
-        "USE \"examples/trek/scan.tbx\"\n",
-        "USE \"examples/trek/combat.tbx\"\n",
-        "USE \"examples/trek/nav.tbx\"\n",
-        "USE \"examples/trek/library.tbx\"\n",
-        "USE \"examples/trek/command.tbx\"\n",
-        "DEF RUN()\n",
-        "  INIT_GAME\n",
-        "  CLEAR_SECTOR\n",
-        "  LET ENT_SX = 4\n",
-        "  LET ENT_SY = 4\n",
-        "  LET @SECTOR[ENT_SX, ENT_SY] = 1\n",
-        "  LET @SECTOR[6, 4] = 3\n",
-        "  LET DOCKED = FALSE\n",
-        "  LET CONDITION = \"GREEN\"\n",
-        "  LET ENERGY = MAX_ENERGY - 100\n",
-        "  LET TORPEDOES = MAX_TORPEDOES - 2\n",
-        "  LET SHIELDS = 250\n",
-        "  LET KLINGONS_HERE = 0\n",
-        "  LET START_STARDATE = 2000\n",
-        "  LET STARDATE = 2000\n",
-        "  LET MISSION_DAYS = 0\n",
-        "  VAR DISCARDED = GET_OUTPUT()\n",
-        "  RUN_COMMAND_LOOP\n",
-        "END\n",
-        "RUN\n",
-    );
-    let (buffered, flushed) =
-        run_trek_src_with_input_and_flushed_output(src, "0\n1\n0.2\n0\n1\n1.0\n");
-    let combined = format!("{flushed}{buffered}");
-    assert_eq!(
-        combined
-            .matches("SHIELDS DROPPED FOR DOCKING PURPOSES\n")
-            .count(),
-        1,
-        "navigation should trigger exactly one docking refresh before the next prompt"
-    );
-}
-
-#[test]
-fn test_trek_command_loop_does_not_double_refresh_after_command_one_scan() {
-    let src = concat!(
-        "USE \"examples/trek/state.tbx\"\n",
-        "USE \"examples/trek/util.tbx\"\n",
-        "USE \"examples/trek/init.tbx\"\n",
-        "USE \"examples/trek/scan.tbx\"\n",
-        "USE \"examples/trek/combat.tbx\"\n",
-        "USE \"examples/trek/nav.tbx\"\n",
-        "USE \"examples/trek/library.tbx\"\n",
-        "USE \"examples/trek/command.tbx\"\n",
-        "DEF RUN()\n",
-        "  INIT_GAME\n",
-        "  CLEAR_SECTOR\n",
-        "  LET ENT_SX = 4\n",
-        "  LET ENT_SY = 4\n",
-        "  LET @SECTOR[ENT_SX, ENT_SY] = 1\n",
-        "  LET @SECTOR[5, 4] = 3\n",
-        "  LET DOCKED = FALSE\n",
-        "  LET CONDITION = \"GREEN\"\n",
-        "  LET ENERGY = MAX_ENERGY - 100\n",
-        "  LET TORPEDOES = MAX_TORPEDOES - 2\n",
-        "  LET SHIELDS = 250\n",
-        "  LET KLINGONS_HERE = 0\n",
-        "  LET START_STARDATE = 2000\n",
-        "  LET STARDATE = 2000\n",
-        "  LET MISSION_DAYS = 0\n",
-        "  VAR DISCARDED = GET_OUTPUT()\n",
-        "  RUN_COMMAND_LOOP\n",
-        "END\n",
-        "RUN\n",
-    );
-    let (buffered, flushed) = run_trek_src_with_input_and_flushed_output(src, "1\n0\n1\n1.0\n");
-    let combined = format!("{flushed}{buffered}");
-    assert_eq!(
-        combined
-            .matches("SHIELDS DROPPED FOR DOCKING PURPOSES\n")
-            .count(),
-        2,
-        "initial scan and command 1 scan should dock once each without an extra post-command refresh"
-    );
-}
-
-#[test]
-fn test_trek_command_loop_does_not_refresh_after_non_navigation_command() {
-    let src = concat!(
-        "USE \"examples/trek/state.tbx\"\n",
-        "USE \"examples/trek/util.tbx\"\n",
-        "USE \"examples/trek/init.tbx\"\n",
-        "USE \"examples/trek/scan.tbx\"\n",
-        "USE \"examples/trek/combat.tbx\"\n",
-        "USE \"examples/trek/nav.tbx\"\n",
-        "USE \"examples/trek/library.tbx\"\n",
-        "USE \"examples/trek/command.tbx\"\n",
-        "DEF RUN()\n",
-        "  INIT_GAME\n",
-        "  CLEAR_SECTOR\n",
-        "  LET ENT_SX = 4\n",
-        "  LET ENT_SY = 4\n",
-        "  LET @SECTOR[ENT_SX, ENT_SY] = 1\n",
-        "  LET @SECTOR[5, 4] = 3\n",
-        "  LET DOCKED = TRUE\n",
-        "  LET CONDITION = \"DOCKED\"\n",
-        "  LET SHIELDS = 200\n",
-        "  LET KLINGONS_HERE = 0\n",
-        "  LET START_STARDATE = 2000\n",
-        "  LET STARDATE = 2000\n",
-        "  LET MISSION_DAYS = 0\n",
-        "  VAR DISCARDED = GET_OUTPUT()\n",
-        "  RUN_COMMAND_LOOP\n",
-        "  PUTDEC SHIELDS\n",
-        "END\n",
-        "RUN\n",
-    );
-    let (buffered, flushed) =
-        run_trek_src_with_input_and_flushed_output(src, "5\n123\n0\n1\n1.0\n");
-    let combined = format!("{flushed}{buffered}");
-    assert_eq!(
-        combined
-            .matches("SHIELDS DROPPED FOR DOCKING PURPOSES\n")
-            .count(),
-        1,
-        "only the initial short-range scan should refresh docking for a non-navigation command cycle"
-    );
-    assert!(
-        combined.ends_with("123"),
-        "shield control value should survive until loop exit without post-command reset"
-    );
-}
-
-#[test]
-fn test_dispatch_command_7_eof_flushes_prompt_and_buffers_menu() {
-    let src = concat!(
-        "USE \"examples/trek/state.tbx\"\n",
-        "USE \"examples/trek/util.tbx\"\n",
-        "USE \"examples/trek/init.tbx\"\n",
-        "USE \"examples/trek/scan.tbx\"\n",
-        "USE \"examples/trek/combat.tbx\"\n",
-        "USE \"examples/trek/nav.tbx\"\n",
-        "USE \"examples/trek/library.tbx\"\n",
-        "USE \"examples/trek/command.tbx\"\n",
-        "DEF RUN()\n",
-        "  INIT_GAME\n",
-        "  VAR DISCARDED = GET_OUTPUT()\n",
-        "  DISPATCH_COMMAND 7\n",
-        "END\n",
-        "RUN\n",
-    );
-    let (buffered, flushed) = run_trek_src_with_input_and_flushed_output(src, "");
-    assert_eq!(flushed, "COMPUTER ACTIVE AND AWAITING COMMAND");
-    assert_eq!(
-        buffered,
-        concat!(
-            "FUNCTIONS AVAILABLE FROM COMPUTER\n",
-            "   0 = CUMULATIVE GALACTIC RECORD\n",
-            "   1 = STATUS REPORT\n",
-            "   2 = PHOTON TORPEDO DATA\n",
-        )
     );
 }
